[ALPS07979684] cam: chunk handle u64
[Detail]
1. sync code from dev branch.

MTK-Commit-Id: 8f569c2d42d8078b6350cf63a6a02ff3aebb7e15

CR-Id: ALPS07979684
Feature: [Android Default] Camera Application Basic Functions

+++ drivers/media/platform/mtk-isp/camsys/isp7s/cam_legacy/mtk_cam-seninf-if.h
@@ -113,7 +113,7 @@ int mtk_cam_seninf_get_pad_data_info(struct v4l2_subdev *sd,
 				struct mtk_seninf_pad_data_info *result);
 
 void
-mtk_cam_seninf_set_secure(struct v4l2_subdev *sd, int enable, unsigned int SecInfo_addr);
+mtk_cam_seninf_set_secure(struct v4l2_subdev *sd, int enable, u64 SecInfo_addr);
 
 bool is_fsync_listening_on_pd(struct v4l2_subdev *sd);
